atomic int++

+++ core.go
@@ -43,7 +43,7 @@ func (core *Core) GenerateBlock() {
 			preHash = newBlock.Header.BlockHash
 			fmt.Printf("Generate new block[%d] tx count= %d, cost: %v, TPS: %v\n", newBlock.Header.BlockHeight,
 				len(txs), time.Since(start), float64(len(txs))/time.Since(start).Seconds())
-			InBlockTxCount += len(txs)
+			InBlockTxCount += uint32(len(txs))
 			start = time.Now()
 		}
 	}()
